Adds addAuthorization function, fixes addQuiz.
There is no need now for a second instance of axios. When the user logs
in, token is stored in local storage. When a request is made that
requires a JSON token, addAuthorization checks local storage and adds
token. When a user logs out, Authorization header is deleted.

I'm thinking there's probably a better way to do this? But this works
okay for now.